machine-learning-client/requirements.txt
breadability==0.1.20
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.5
dnspython==2.7.0
docopt==0.6.2
emoji==1.7.0
exceptiongroup==1.2.2
gensim==4.3.3
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
lxml==5.3.0
nltk==3.8.1
numpy==1.26.4
packaging==24.2
pluggy==1.5.0
pycountry==24.6.1
pymongo==4.10.1
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
scipy==1.13.1
smart-open==7.0.5
sumy==0.11.0
text2emotion==0.0.5
textblob==0.18.0.post0
tomli==2.1.0
tqdm==4.67.0
urllib3==2.2.3
vaderSentiment==3.3.2
wrapt==1.16.0